+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Requires macOS 11.3 or newer, for 11.2.3 and older use Minimal or higher spoofing
   - See additional notes before updating: [VMM usage notes](https://github.com/dortania/OpenCore-Legacy-Patcher/issues/543#issuecomment-953441283)
 - Adjust SIP setting to better reflect current SIP usage
+- Resolve Monterey Bluetooth issues on user-upgraded BCM94331 BT4.0 modules
 
 ## 0.3.1
 - Increment Binaries: